+++ src/metafiles/mod.rs
@@ -1,3 +1,4 @@
 pub mod config_file;
 pub mod project_file;
 pub mod registry_file;
+pub mod trust_store;